Adjust detected face boundaries with video frame and video output
Previously detected face boundaries were not considering boundaries of
an output video image, thus Label was drawn somewhere nearby the face.
This was replaced with a Rectangle and correct position was adjusted
with accordance to current boundaries of VideoOutput - it could be
resized and moved, but the Rectangle will still be drawn around the
face.